+++ previous-versions/assess/RELEASES.md
@@ -667,3 +667,13 @@ Based on:
 - [typescript v0.37.1] previous-versions/assess
 ### Releases
 - [NPM v0.37.1] https://www.npmjs.com/package/@codat/assess/v/0.37.1 - previous-versions/assess
+
+## 2023-10-03 00:20:50
+### Changes
+Based on:
+- OpenAPI Doc 1.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Assess.yaml
+- Speakeasy CLI 1.93.1 (2.144.7) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.37.2] previous-versions/assess
+### Releases
+- [NPM v0.37.2] https://www.npmjs.com/package/@codat/assess/v/0.37.2 - previous-versions/assess